decreased the load time by multi threading
Decreased the load time by changing the loops of building frames with multi threading(Parallel.For) .
Also added keeping the aspect ratio of the video and the output same.
And user can now choose 2 different type of words that will be used in output, one is better for lower resolution, one is better for high.
You have to put ffprobe.exe in your folder now(for keeping the aspect ratio same).